Moved Location Search to its own component
